skeleton: finish reading chapter 7 - Translation
So hard...

Add a lot of comments and hints based on the textbook and my naive understanding.

Classes with all its members public-ed are changed to be structures, which are more naturally.

TODO: should I modify some poorly-named fields? For example, the field `tr::Cx::stm_` is public,
but suffixed by an underscore, which is not intuitive.