fix: x86 stack pointer
X86 specifies that the register RSP should keep the stack pointer.
When the stack grows/shrinks because of PUSH/POP, RSP correspondingly decreases/increases.

Meanwhile, X86 has the **frame pointer** (a.k.a. base pointer) - RBP,
which serves as a _snapshot_ of the stack pointer.
The frame pointer offers the ability of dynamically changing the frame size.

However, in this lab, the frame pointer is not allowed to use.
It's like that we enable the optimization option `-fomit-frame-pointer` in GCC.
As a result, there's no need to maintain the frame pointer RBP.
So the calling convention can be simplifies.