fix: munch is guaranteed to return fresh register
The interface Exp#Munch returns temp::Temp, which is a register in the real world.
It should always allocate a new register and transfer the data into it,
and return the register.

The benefits:
- It's always safe to modify the register, as it's newly-allocated. No conflicts occurs when modifying it.
- The behavior can be consistent. Without reading the code, the programmer has the confidence that
  he or she can always return the return value of any Munch as another return value of Munch.
- No factual effects for the code generation. Later we will do register allocation, and the transient registers
  allocated by this will be eliminated.

It's very common to generate code like this:

```
movq %rax, t118
movq t118, t116
... ; use t116 to do more things
```

And the register allocator can easily discover the redundance of t118, therefore eliminate it.