fix(escape): compromise for Lab3
The commit does not make any sense. It should not be applied.

Previously, I change the constructor of abstract syntax tree nodes that have `escape_` field.
I initialize `escape_` to `false`, which is a save for the frame size.

But the modification causes Lab3 test fails, as it unexpectedly check the `escape_` field.
In Lab3 (test_parse.cc), the escape analysis isn't done. So the compiler has no idea about which variable is escaped.

It's better to resolve this paradox by modifying the test script. I've also commit a patch.