fix: how to check "precolored" temporaries?
In summary, before the register allocator is constructed, we must use the global variable `reg_manager`; after the register allocator is fully constructed (that is, `RegAllocator::RegAllocator` is called), we can kindly use the member function `RegAllocator::is_precolored`, which just check the set `precolored`.

Note that the set `precolored` is unchanged since the register allocator is constructed (because the program rewriting caused by a real spill won't add new precolored temporaries). Therefore, it's always safe to use `RegAllocator::is_precolored` after that.